feat: GetTimeGMT() function
